+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v0.1.3:
+  * Implement reg_write and action instruction.
   * Implement custom raw instructions.
   * Include correct license file.